backend/german-agent-microservice/requirements.txt
accelerate==0.29.3
boto3==1.34.101
deepl==1.18.0
elevenlabs==0.2.27
Flask==3.0.3
Flask-API==3.1
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_experimental==0.0.58
langchain_openai==0.1.6
langchain_weaviate==0.0.1.post1
llama_parse==0.4.2
Markdown==3.6
openai==0.28.0
python-dotenv==1.0.1
rank-bm25==0.2.2
sentence-transformers==2.7.0
simsimd==3.7.7
tavily-python==0.3.3
tokenizers==0.19.1
transformers==4.40.1
weaviate_client==4.5.7
unstructured==0.13.7